+++ beam-postgres/test/Main.hs
@@ -11,6 +11,7 @@ import qualified Database.Beam.Postgres.Test.Select as Select
 import qualified Database.Beam.Postgres.Test.Marshal as Marshal
 import qualified Database.Beam.Postgres.Test.DataTypes as DataType
 import qualified Database.Beam.Postgres.Test.Migrate as Migrate
+import qualified Database.Beam.Postgres.Test.TempTable as TempTable
 import Database.PostgreSQL.Simple ( ConnectInfo(..), defaultConnectInfo )
 import qualified Database.PostgreSQL.Simple as Postgres
 
@@ -23,6 +24,7 @@ main = defaultMain
           , Select.tests getConnStr
           , DataType.tests getConnStr
           , Migrate.tests getConnStr
+          , TempTable.tests getConnStr
           ]